feat: created a navbar and footer component
